properly cast sensor values

+++ Software/Projects/B-L072Z-LRWAN1/Applications/LoRa/End_Node/LoRaWAN/App/src/bsp.c
@@ -137,11 +137,11 @@ void BSP_sensor_Read(void)
 	
 	
 	sensor_data.humidity    = 34; // hard coded dummy value
-	sensor_data.temperature = TEMPERATURE_Value;
-	sensor_data.pressure    = PRESSURE_Value;
-	sensor_data.no_load_solar_voltage = no_load_solar_voltage/100;
-	sensor_data.load_solar_voltage = load_solar_voltage/100;
-	sensor_data.sats = gps_info.GPSsats;
+	sensor_data.temperature = (int8_t)TEMPERATURE_Value;
+	sensor_data.pressure    = (uint16_t)PRESSURE_Value;
+	sensor_data.no_load_solar_voltage = (uint8_t)(no_load_solar_voltage/100);
+	sensor_data.load_solar_voltage = (uint8_t)(load_solar_voltage/100);
+	sensor_data.sats = (uint8_t)gps_info.GPSsats;
 
 
 	current_position.altitude  = (gps_info.GPSaltitude >> 8) & 0xffff;
